Back/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-framework-nested==0.0.1
djangorestframework==3.11.0
drf-nested-routers==0.91
et-xmlfile==1.1.0
idna==3.2
jdcal==1.4.1
openpyxl==3.0.3
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.25.8
webdriver-manager==3.4.2
